Move and delete InpregHandler
